Reorganize VariableType package and add package-info files
