+++ src/main/java/zowe/client/sdk/zosvariables/input/factory/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Input data classes and factory methods for z/OSMF system variables.
+ */
+package zowe.client.sdk.zosvariables.input.factory;
