+++ src/main/java/zowe/client/sdk/zosvariables/response/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * z/OSMF system variables response models.
+ */
+package zowe.client.sdk.zosvariables.response;
