+++ src/main/java/zowe/client/sdk/zosvariables/type/VariableType.java
@@ -7,7 +7,7 @@
  *
  * Copyright Contributors to the Zowe Project.
  */
-package zowe.client.sdk.zosvariables.input.factory;
+package zowe.client.sdk.zosvariables.type;
 
 /**
  * Supported variable source types.
@@ -30,6 +30,7 @@ public enum VariableType {
     /**
      * REST API value.
      */
+
     private final String value;
     VariableType(final String value) {
         this.value = value;
@@ -40,6 +41,7 @@ public enum VariableType {
      *
      * @return value
      */
+
     public String getValue() {
         return value;
     }
