+++ src/main/java/zowe/client/sdk/zosvariables/type/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Enumerations for z/OSMF system variables.
+ */
+package zowe.client.sdk.zosvariables.type;
