+++ src/test/java/zowe/client/sdk/zosvariables/input/factory/VariableGetInputDataTest.java
@@ -10,6 +10,8 @@
 package zowe.client.sdk.zosvariables.input.factory;
 
 import org.junit.jupiter.api.Test;
+import zowe.client.sdk.zosvariables.type.VariableType;
+
 import java.util.Arrays;
 import static org.junit.jupiter.api.Assertions.*;
