+++ src/test/java/zowe/client/sdk/zosvariables/input/factory/VariableGetInputFactoryTest.java
@@ -9,6 +9,7 @@
  */
 package zowe.client.sdk.zosvariables.input.factory;
 import org.junit.jupiter.api.Test;
+import zowe.client.sdk.zosvariables.type.VariableType;
 import java.util.Arrays;
 import static org.junit.jupiter.api.Assertions.*;
